renamed helpers file, excluded it from tests
