+++ integration/.env
@@ -1,5 +1,5 @@
-RPC_URL=
-CHAIN_ID=
-ACCOUNT_PRIVATE_KEY=
-HANJI_API_BASE_URL=
-HANJI_WEBSOCKET_BASE_URL=
+RPC_URL=https://node.ghostnet.etherlink.com
+CHAIN_ID=128123
+ACCOUNT_PRIVATE_KEY=cb2b1e236822479d8c71b5bc054ceafc4275833855e46f43d1c0b661a91d8116
+HANJI_API_BASE_URL=https://api-dev.hanji.io
+HANJI_WEBSOCKET_BASE_URL=wss://sockets-dev.hanji.io
