+++ jest.config.ts
@@ -11,5 +11,6 @@ const jestConfig: JestConfigWithTsJest = {
       },
     ],
   },
+  testPathIgnorePatterns: ['/src/helpers.test.ts']
 };
 export default jestConfig;
